+++ tests/test_response_context.py
@@ -3,7 +3,7 @@ import pytest
 from falcon import Response
 
 
-class TestRequestContext(object):
+class TestResponseContext(object):
 
     def test_default_response_context(self):
         resp = Response()
@@ -11,6 +11,11 @@ class TestRequestContext(object):
 
         resp.context.hello = 'World!'
         assert resp.context.hello == 'World!'
+        assert 'hello' not in resp.context
+
+        resp.context['note'] = 'Default Response.context_type used to be dict.'
+        assert 'note' in resp.context
+        assert resp.context.get('note') == resp.context['note']
 
     def test_custom_response_context(self):